feat: added loading page andd not found
