+++ app/search/not-found.tsx
@@ -1,0 +1,16 @@
+import Image from 'next/image';
+export default function NotFound() {
+
+  return (
+    <main className="flex items-center justify-center flex-col flex-1">
+      <Image
+        src="/no-pokemon-found.png"
+        height={300}
+        width={300}
+        alt="loading picachu"
+        className="h-70 w-70"
+      ></Image>
+      <p className="text-4xl font-bold mt-4">No pokemon found</p>
+    </main>
+  );
+}
