+++ components/search-input.tsx
@@ -1,6 +1,6 @@
 import { Input } from '@/components/ui/input';
 import Form from 'next/form';
-export default function Search({ query }: { query?: string }) {
+export default function SearchInput({ query }: { query?: string }) {
   return (
     <Form className="flex justify-center py-10" action="/search">
       <Input
